fix long type

+++ in_progress/push_swap/utils_calc_moves.c
@@ -6,7 +6,7 @@
 /*   By: achauvie <achauvie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/18 09:16:00 by achauvie          #+#    #+#             */
-/*   Updated: 2025/11/18 13:41:07 by achauvie         ###   ########.fr       */
+/*   Updated: 2025/11/21 14:22:13 by achauvie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ size_t	find_stack_max_pos(t_stack *stack)
 {
 	size_t	i;
 	size_t	tmp_i;
-	size_t	tmp_max;
+	long	tmp_max;
 
 	tmp_max = stack->nbr;
 	i = 0;
@@ -36,7 +36,7 @@ size_t	find_stack_max_pos(t_stack *stack)
 
 long	find_stack_max_value(t_stack *stack)
 {
-	size_t	tmp_max;
+	long	tmp_max;
 
 	tmp_max = stack->nbr;
 	while (stack)
@@ -52,7 +52,7 @@ size_t	find_stack_min_pos(t_stack *stack)
 {
 	size_t	i;
 	size_t	tmp_i;
-	size_t	tmp_min;
+	long	tmp_min;
 
 	tmp_min = stack->nbr;
 	i = 0;
@@ -72,7 +72,7 @@ size_t	find_stack_min_pos(t_stack *stack)
 
 long	find_stack_min_value(t_stack *stack)
 {
-	size_t	tmp_min;
+	long	tmp_min;
 
 	tmp_min = stack->nbr;
 	while (stack)
